+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## v2.0.15
+FEATURES:
+* Get/set device Scheduled Bells
+
 IMPROVEMENTS:
 * `RelayMode` is now an enum
 * Check TimeZone ranges